pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.67.0"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
SQLAlchemy = "^1.4.22"
asyncpg = "^0.25.0"
alembic = "^1.7.7"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.15.1"
aiosqlite = "^0.17.0"
httpx = "^0.18.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
